requirements.txt
adbs==0.0.2
asgiref==3.2.7
attrs==21.2.0
Automat==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2021.5.30
cffi==1.14.0
chardet==3.0.4
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.0.5
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-crispy-forms==1.9.0
django-filter==2.4.0
django-jinja==2.7.0
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-nepali-datetime-field==0.4
django-phone-field==1.8.1
django-postgres-extensions==0.9.3
django-rest-framework==0.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djongo==1.3.6
google-api-core==1.30.0
google-auth==1.32.0
google-cloud-core==1.7.1
google-cloud-translate==3.2.1
google-trans-new==1.1.9
googleapis-common-protos==1.53.0
googletrans==3.0.0
grpcio==1.38.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
html5lib==1.0.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==4.5.0
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
lxml==4.6.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
nepali==0.4.2
nepali-datetime==1.0.6
nltk==3.6.2
numpy==1.20.3
packaging==20.9
pandas==1.2.4
parsel==1.6.0
Pillow==7.1.1
priority==1.3.0
Protego==0.1.16
proto-plus==1.18.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyBSDate==0.3.0
pycparser==2.20
PyDispatcher==2.0.5
pymongo==3.12.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
Pyphen==0.9.5
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
pyshorteners==1.0.1
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2019.3
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
queuelib==1.6.1
regex==2021.4.4
requests==2.25.1
rfc3986==1.5.0
rsa==4.7.2
Scrapy==2.5.0
screeninfo==0.6.7
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
sqlparse==0.2.4
textblob==0.15.3
tinycss2==1.0.2
tqdm==4.61.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
typing-extensions==3.10.0.0
urllib3==1.26.5
w3lib==1.22.0
WeasyPrint==51
webencodings==0.5.1
zipp==3.4.1
zope.interface==5.4.0
